wip(backend): completed first part of creating order.

+++ Byte&Buy.Backend/src/ByteBuy.Core/Domain/Entities/RentOrderLine.cs
@@ -29,7 +29,7 @@ public class RentOrderLine : OrderLine
         string imagePath,
         string? altText,
         int quantity,
-        int amount,
+        decimal amount,
         string currency,
         int rentalDays)
     {
